Fix logic of timeout writing nulls
